Add a Living Module
This module can notify a process when its alive. This makes it so that
we can wrap a child module with a notification message telling the
above process that this one is back online.

Without Misc.Alive, one would have to use a timer to hope the process
has came back online