Start an Inspector Application
This is the first commit of the new inspector tool:

The only view we currently support is raw, which just lists the slots
in a record.

Currently we do not extend it with any additional functionality.

However this code will be ported to Elixir and use some protocols
define to better integrate many different ideas.