Make the Raw View reucrisvely display nested items
Instead of baking in the display message, we instead make a proper
`raw/3` function that recursively displays the item.

We also contemplate not using the `view/1` as it's written as one can
not display objects, however this view was specifically progressed as
text could maybe be "good enough" for this particular view.

Further we make sure to make sure the columns properly represent the
data themselves